backend/requirements.txt
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.6.0
pycryptodome==3.9.9
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1